Added Basic Implementation of user.Service.FindUsers ==================================================== * Added Query and Page types to userstore * Addressed issue with types used for page and total in query and page types * Added FindMany to users.UserStore service * Added necessary stubs to stubUserStore * Added basic implementation of user.Service.FindUsers plus happy path tests

+++ pkg/rpc/rpc_test.go
@@ -181,7 +181,7 @@ func usersPageFromQuery(query user.Query) user.Page {
 	}
 	return user.Page{
 		Page:  query.Page,
-		Total: query.Page * query.Length,
+		Total: query.Page * int64(query.Length),
 		Items: items,
 	}
 }
